Split out RealtimeSync into separate section
